+++ CHANGELOG.md
@@ -5,6 +5,7 @@
   * Introduced the ``CP2KMM`` class for classical forcefield calculations with CP2K: [qmflows/pull/150](https://github.com/SCM-NV/qmflows/pull/150).
   * Introduced the ``PackageWrapper`` class: [qmflows/pull/149](https://github.com/SCM-NV/qmflows/pull/149).
   * Introduced updates and code-style improvements to the ``Package`` and ``Result`` classes: [qmflows/pull/146](https://github.com/SCM-NV/qmflows/pull/146).
+  * [Removed references to Dirac](https://github.com/SCM-NV/qmflows/issues/152)
 
 
 # Version 0.9.0 (27/11/2019)